backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
psycopg2-binary==2.9.9
python-decouple==3.8
requests==2.31.0
pillow==10.1.0
django-filter==23.3
drf-yasg==1.21.7
celery==5.3.4
redis==5.0.1
python-jose==3.3.0
cryptography==41.0.7
